+++ app_init.py
@@ -1,0 +1,5 @@
+from app import init_db_and_create_first_user
+
+
+if __name__ == '__main__':
+    init_db_and_create_first_user()